Python/Image-Viewer/requirements.txt
python-resize-image==1.1.11
pillow==4.3.0
tkintertable==1.3.2